Terrorism Analysis & Finding Insights
In the present era, the data generated is doubling every 24 months. The area of data analysis and representation has become a necessity for interpreting large amounts of data for having a clear insight.
In this project, we got a client problem, where a client approached, for having a dataset for Terrorism Analysis & Finding its Insights. The client explained the problems, and also described all the parameters he needed from us for having a clear insight.